added starter hello world for app
