+++ .gitignore
@@ -1,6 +1,7 @@
 # OSX
 #
 .DS_Store
+.vscode
 
 # Xcode
 #
