+++ README.md
@@ -10,9 +10,11 @@
 | Vincent W | developer + pm | @wuv21
 
 ## Setup
-These are the instructions to set up this repo for editing.
+These are the instructions to set up this repo for editing. These steps require using macOS for development.
 
-1. To be added
+1. Install node via `brew install node`. Check that it is installed by typing `node --version`.
+2. Install Xcode via Apple's developer tools or app store.
+3. TBA
 
 ## References
 - TBA
